add ability to override pgo path

+++ src/lib.rs
@@ -94,7 +94,9 @@ fn get_rustc_info(field: &str) -> anyhow::Result<String> {
 
 /// Clears all files from the directory, and recreates it.
 fn clear_directory(path: &Path) -> std::io::Result<()> {
-    std::fs::remove_dir_all(path)?;
+    if path.exists() {
+        std::fs::remove_dir_all(path)?;
+    }
     ensure_directory(path)
 }
